Convert MoveLambdaOutsideParenthesesIntention to MoveLambdaOutsideParenthesesInspection
