+++ idea/resources/inspectionDescriptions/MoveLambdaOutsideParentheses.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+This inspection moves a lambda expression outside parentheses
+</body>
+</html>
