--- idea/resources/inspectionDescriptions/RemoveRedundantConstructorParentheses.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-This inspection reports redundant parentheses in constructor.
-</body>
-</html>
