--- idea/resources/intentionDescriptions/MoveLambdaOutsideParenthesesIntention/description.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-This intention moves a lambda expression outside parentheses.
-</body>
-</html>
