+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/.inspection
@@ -1,0 +1,1 @@
+org.jetbrains.kotlin.idea.inspections.MoveLambdaOutsideParenthesesInspection
