+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/ambigousOverload.kt.after
