+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/functionalValueCall.kt
