+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/inapplicable1.kt
@@ -1,4 +1,4 @@
-// IS_APPLICABLE: false
+// PROBLEM: none
 fun foo() {
     bar() <caret>{ it }
 }
