+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/labeledLambda.kt.after
