+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/lambdaWithCommas.kt.after
