+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/lambdaWithCommas2.kt
