+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/lambdaWithCommas3.kt
