+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda1.kt
