+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda2.kt
