+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda3.kt
