+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda4.kt
@@ -3,7 +3,7 @@ fun foo() {
     bar(2, {
         val x = 3
         it * x
-    })<caret>
+    <caret>})
 }
 
 fun bar(a: Int, b: (Int) -> Int) {
