+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda4.kt.after
