+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda5.kt.after
