+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda6.kt
@@ -1,6 +1,6 @@
 // IS_APPLICABLE: true
 fun foo() {
-    bar(3, 2, 1, { it }<caret>)
+    bar(3, 2, 1, { it <caret>})
 }
 
 fun bar(name1: Int, name2: Int, name3: Int, name4: (Int) -> Int) : Int {
