+++ idea/testData/inspectionsLocal/moveLambdaOutsideParentheses/moveLambda6.kt.after
