--- idea/testData/inspectionsLocal/removeRedundantConstructorParentheses/.inspection
@@ -1,1 +1,0 @@
-org.jetbrains.kotlin.idea.inspections.RemoveRedundantConstructorParenthesesInspection
