--- idea/testData/inspectionsLocal/removeRedundantConstructorParentheses/basic.kt
@@ -1,6 +1,0 @@
-// WITH_RUNTIME
-class MyClass(val function: (Int) -> Boolean)
-
-fun foo() {
-    MyClass(<caret>{ it == 7 })
-}
