--- idea/testData/inspectionsLocal/removeRedundantConstructorParentheses/noFunction.kt
@@ -1,6 +1,0 @@
-// PROBLEM: none
-class MyClass(val x: Int)
-
-fun foo() {
-    MyClass(<caret>7)
-}
