--- idea/testData/intentions/moveLambdaOutsideParentheses/.intention
@@ -1,1 +1,0 @@
-org.jetbrains.kotlin.idea.intentions.MoveLambdaOutsideParenthesesIntention
